敵とプレイヤーのスロー同期、ジャスト回避処理固め
06/02＿スローの処理をプレイヤーとエネミーに同期させた。
エネミー側にジャスト回避可能フラグがありプレイヤーは参照して可能ならジャストを行う

+++ san_framework/public/object/character.cpp
@@ -252,7 +252,7 @@ int cCharacter::getPartsNum(void)
 	return PartsNum;
 }
 
-void cCharacter::setAnimSpeed(int speed)
+void cCharacter::setAnimSpeed(float speed)
 {
 	animSpeed = speed;
 }
